# Improve bios memspeed performances